Hack: Implement Realtime Audio Playback Option
From the Lime3DS PR:

This pull request ports the realtime audio option which was recently added to Citra Enhanced, although I have made a notable improvement to how the feature behaves.

The version of the feature which is currently present in Citra Enhanced calculates the audio timescale by using only data from the previous frame. This means that if the emulation framerate is inconsistent, the audio playback will reflect this inconsistency, leading to frequent and very noticeable speedups and slowdowns.

To remedy this, I have tweaked the implementation of this feature when bringing it over to Lime3DS. Instead of only using the data from the previous frame to calculate the audio timescale, a mean average of the last 50 displayed frames is used. This results in significantly smoother playback when the framerate is inconsistent.

This '50 frames' number was the result of personal testing. Values lower than this did not properly address the issue enough and still had speedups and slowdowns in audio playback when framerate was inconsistent, whereas values higher than this did not account for changes in framerate fast enough, which resulted in the same issue when entering large lag spikes. 50 frames of data seems to be the Goldilocks zone for consistent playback.

Squashed commit of the following:

commit e0ae4a6700b245fbf8a4559f0fdc44e7e5f840d9
Author: OpenSauce04 <opensauce04@gmail.com>
Date:   Thu May 16 22:34:13 2024 -0600

    Desktop: HLE audio settings are greyed out when non-HLE audio emulation is selected

commit 141d9764682234db4b28c917785db62cda873fdc
Author: OpenSauce04 <opensauce04@gmail.com>
Date:   Thu May 16 20:00:39 2024 +0000

    Fixed Windows build issues

commit a1ea21cd51cbae31782fbf53d5aa423acef67b65
Author: OpenSauce04 <opensauce04@gmail.com>
Date:   Thu May 16 16:20:39 2024 +0100

    Updated name and description of realtime audio setting

commit 25cc6c9d2054e8ad76d16cc9aacc37f45d93be66
Author: OpenSauce04 <opensauce04@gmail.com>
Date:   Thu May 16 15:53:13 2024 +0100

    Improve playback consistency for realtime audio during inconsistent framerates

    Implements and uses `GetStableFrameTimeScale` instead of `GetLastFrameTimeScale` for realtime audio playback