Basic implementation of AC::ScanAPs and NWM_INF::RecvBeaconBroadcastData
Experimental.

Ported from Lime3DS.

From the original PR:

What: This is a basic implementation of the unimplemented methods ScanAPs and RecvBeaconBroadcastData from Service::AC::Module::Interface and Service::NWM::NWM_INF respectively. It is as of yet not fully finished, but I hope it will serve as a foundation for later modifications. The purpose of ScanAPs is to give a list of data on nearby networks to the game.

Why: The reason I tried implementing ScanAPs was because it is a feature a few games rely on to function properly. Most notably, the Denpa Men games rely on this to discover nearby networks. Without this method, you would need a save file editor or a pre-existing save file to even play the game, which takes most of the fun away from it.

    How: I mostly went on what little information is available on this method on 3dbrew.org. Based on that, I tried to construct ScanAPs so that it would first call NWM_INF::RecvBeaconBroadcastData and pass on the raw beacon data to its own caller, as the parsing method is yet unknown. Since NWM_INF::RecvBeaconBroadcastData was also unimplemented and had a nearly identical use as NWM_UDS::RecvBeaconBroadcastData, I had it delegate the task to NWM_UDS::RecvBeaconBroadcastData by reconstructing the command buffer again. I did attempt to do an independent implementation, but it required adding an additional private method to `NWM_INF`` that was not mentioned in the header file or anywhere else, so I decided to discard the idea for now.

Result: In the release version, entering the Denpa tower would result in the main Denpa telling you no radio waves could be found. After this change, this does not differ. However, the log now indicates that it might be an issue with writing to the MappedBuffer instance during the NWM_UDS::RecvBeaconBroadcastData, instead of telling you that the method is unimplemented. This could imply that in the way I called the other services, something goes wrong with the mapped buffer or the mapped buffer is not properly retrieved during ScanAPs.